+++ sc-memory/cpp/sc_template_search.cpp
@@ -27,70 +27,132 @@ public:
 
   void UpdateSearchCache()
   {
-    if (!m_template.IsSearchCacheValid())
+    if (!m_template.IsSearchCacheValid() && !m_template.m_constructions.empty())
     {
       // update it
       auto const & triples = m_template.m_constructions;
-      ScTemplate::ProcessOrder & cache = m_template.m_searchCachedOrder;
-      cache.resize(m_template.m_constructions.size());
-      size_t i = 0;
-      for (size_t & v : cache)
-        v = i++;
+      ScTemplate::ProcessOrder preCache(m_template.m_constructions.size());
+      for (size_t i = 0; i < preCache.size(); ++i)
+        preCache[i] = i;
 
-      /* Main idea of this stage, to build an order where iterators with maximum number of
-       * fixed elements will runs first. This allow to minimize number of iterations. Also we can't
-       * start search from template a_a_a, or any template where all fixed elements are remplacements
-       */
-      std::sort(cache.begin(), cache.end(), [&triples](size_t a, size_t b) {
-        ScTemplateConstr3 const & aTriple = triples[a];
-        ScTemplateConstr3 const & bTriple = triples[b];
+      static const size_t kScoreEdge = 5;
+      static const size_t kScoreOther = 1;
 
-        // compare by addrs arguments count
-        size_t const aAddrCount = aTriple.CountAddrs();
-        size_t const bAddrCount = bTriple.CountAddrs();
-
-        if (aAddrCount != bAddrCount)
-          return (aAddrCount > bAddrCount);
-
-        // compare by fixed arguments count
-        size_t const aFCount = aTriple.CountFixed();
-        size_t const bFCount = bTriple.CountFixed();
-
-        if (aFCount != bFCount)
-          return (aFCount > bFCount);
-
-        // compare replacement types
-        auto const CalculateReplTypeValue = [](ScTemplateConstr3 const & tr)
+      /** First of all we need to calculate scores for all triples 
+        * (more scores - should be search first).
+        * Also need to store all replacements that need to be resolved
+        */
+      std::vector<uint8_t> tripleScores(m_template.m_constructions.size());
+      std::unordered_map<std::string, std::vector<size_t>> replDependMap;
+      for (size_t i = 0; i < m_template.m_constructions.size(); ++i)
+      {
+        ScTemplateConstr3 const & triple = m_template.m_constructions[i];
+        auto const CalculateScore = [](ScTemplateConstr3 const & constr)
         {
-          size_t res = 0;
-          for (size_t i = 0; i < 3; ++i)
-          {
-            auto const & v = tr.GetValues()[i];
-            if (v.m_itemType == ScTemplateItemValue::Type::Replace)
-              ++res;
-          }
-          return res;
+          uint8_t score = 0;
+          auto const & values = constr.GetValues();
+          if (values[1].IsAddr() && values[0].IsAssign() && values[2].IsAssign())
+            score += kScoreEdge;
+          else if (values[0].IsAddr() && values[1].IsAssign() && values[2].IsAddr())
+            score += kScoreOther * 2; // should be a sum of (f_a_a and a_a_f)
+          else if (values[0].IsAddr() || values[2].IsAddr())
+            score += kScoreOther;
+
+          return score;
+        };
+        tripleScores[i] = CalculateScore(triple);
+        // doesn't add edges into depend map
+        auto const TryAppendRepl = [&](ScTemplateItemValue const & value, size_t idx)
+        {
+          SC_ASSERT(idx < 3, ());
+          if (!value.IsAddr() && !value.m_replacementName.empty())
+            replDependMap[value.m_replacementName].push_back((i << 2) + idx);
         };
 
-        size_t const aRV = CalculateReplTypeValue(aTriple);
-        size_t const bRV = CalculateReplTypeValue(bTriple);
+        // do not use loop, to make it faster (not all compilators will make a linear code)
+        TryAppendRepl(triple.GetValues()[0], 0);
+        TryAppendRepl(triple.GetValues()[1], 1);
+        TryAppendRepl(triple.GetValues()[2], 2);
+      }
 
-        if (aRV != bRV)
-          return (aRV < bRV);
-
-        // compare by replacements count
-        size_t const aRCount = aTriple.CountReplacements();
-        size_t const bRCount = bTriple.CountReplacements();
-
-        if (aRCount != bRCount)
-          return (aRCount > bRCount);
-
-        auto const & aValues = aTriple.GetValues();
-        auto const & bValues = bTriple.GetValues();
-
-        return false;
+      // sort by scores
+      std::sort(preCache.begin(), preCache.end(), [&](size_t a, size_t b)
+      {
+        return (tripleScores[a] > tripleScores[b]);
       });
+
+      // now we need to append triples, in order, when previous resolve replacement for a next one
+      ScTemplate::ProcessOrder & cache = m_template.m_searchCachedOrder;
+      cache.resize(preCache.size());
+
+      std::vector<bool> isTripleCached(cache.size(), false);
+      std::unordered_set<std::string> usedReplacements;
+
+      size_t preCacheIdx = 1;
+      size_t orderIdx = 1;
+      cache[0] = preCache[0];
+      isTripleCached[cache[0]] = true;
+      while (orderIdx < cache.size())
+      {
+        size_t const curTripleIdx = cache[orderIdx - 1];
+        auto const & triple = m_template.m_constructions[curTripleIdx];
+
+        // get resolved replacements by the last triple
+        std::vector<std::string> resolvedReplacements;
+        resolvedReplacements.reserve(3);     
+        for (auto const & value : triple.GetValues())
+        {
+          if (!value.m_replacementName.empty())
+            resolvedReplacements.emplace_back(value.m_replacementName);
+        }
+
+        // collect triples, that depend on resolved replacements and update their scores
+        std::unordered_set<size_t> resolvedTriples;
+        for (auto const & name : resolvedReplacements)
+        {
+          if (usedReplacements.find(name) != usedReplacements.end())
+            continue;
+
+          auto const it = replDependMap.find(name);
+          if (it != replDependMap.end())
+            resolvedTriples.insert(it->second.begin(), it->second.end());
+
+          usedReplacements.insert(name);
+        }
+
+        // find next non used triple
+        while (isTripleCached[preCache[preCacheIdx]])
+          preCacheIdx++;
+
+        size_t bestTripleIdx = preCacheIdx;
+        // now update scores of resolved triples and find best one scores
+        for (size_t idx : resolvedTriples)
+        {
+          // unpack it
+          size_t const tripleIdx = idx >> 2;
+          size_t const elIdx = idx & 3;
+
+          if (isTripleCached[tripleIdx])
+            continue;
+
+          // for a constants see initial calculation
+          if (idx == 1)
+            tripleScores[tripleIdx] += kScoreEdge;
+          else
+            tripleScores[tripleIdx] += kScoreOther;
+
+          // check if it better
+          if (tripleScores[bestTripleIdx] < tripleScores[tripleIdx])
+            bestTripleIdx = tripleIdx;
+        }
+
+        // append new best triple
+        cache[orderIdx++] = bestTripleIdx;
+        isTripleCached[bestTripleIdx] = true;
+      }
     }
+
+    m_template.m_isSearchCacheValid = true;
   }
 
   ScAddr const & ResolveAddr(ScTemplateItemValue const & value) const
@@ -130,7 +192,7 @@ public:
 
   ScIterator3Ptr CreateIterator(ScTemplateConstr3 const & constr)
   {
-    ScTemplateItemValue const * values = constr.GetValues();
+    auto const & values = constr.GetValues();
 
     ScTemplateItemValue const & value0 = values[0];
     ScTemplateItemValue const & value1 = values[1];
@@ -148,7 +210,7 @@ public:
       return type;
     };
 
-    if (addr0.IsValid())
+    if (addr0.IsValid() && !addr1.IsValid())
     {
       if (addr2.IsValid()) // F_A_F
       {
@@ -159,18 +221,17 @@ public:
         return m_context.Iterator3(addr0, PrepareType(value1.m_typeValue), PrepareType(value2.m_typeValue));
       }
     }
-    else if (addr2.IsValid()) // A_A_F
+    else if (addr2.IsValid() && !addr0.IsValid()) // A_A_F
     {
       return m_context.Iterator3(PrepareType(value0.m_typeValue), PrepareType(value1.m_typeValue), addr2);
     }
-    else if (addr1.IsValid()) // A_F_A
+    else if (addr1.IsValid() && !addr0.IsValid() && !addr2.IsValid()) // A_F_A
     {
       return m_context.Iterator3(PrepareType(value0.m_typeValue), addr1, PrepareType(value2.m_typeValue));
     }
-    else // unknown iterator type
-    {
-      SC_THROW_EXCEPTION(utils::ExceptionInvalidState, "Unknown iterator type");
-    }
+   
+    //// unknown iterator type
+    //SC_THROW_EXCEPTION(utils::ExceptionInvalidState, "Unknown iterator type");
 
     return ScIterator3Ptr();
   }
@@ -215,6 +276,59 @@ public:
     }
   }
 
+  template <typename FuncT>
+  void ForEachIteration(ScTemplateConstr3 const & constr, FuncT && fn)
+  {
+    // try with iterator search
+    ScIterator3Ptr const it3 = CreateIterator(constr);
+    if (it3.IsPtrValid())
+    {
+      while (it3->Next())
+      {
+        /// check if search in structure
+        if (m_struct.IsValid())
+        {
+          if (!CheckInStruct(it3->Get(0)) ||
+              !CheckInStruct(it3->Get(1)) ||
+              !CheckInStruct(it3->Get(2)))
+          {
+            continue;
+          }
+        }
+
+        fn(it3->Get(0), it3->Get(1), it3->Get(2));
+      }
+    }
+    else // try special checks and search
+    {
+      auto const & values = constr.GetValues();
+      ScTemplateItemValue const & value0 = values[0];
+      ScTemplateItemValue const & value1 = values[1];
+      ScTemplateItemValue const & value2 = values[2];
+
+      ScAddr const addr0 = ResolveAddr(value0);
+      ScAddr const addr1 = ResolveAddr(value1);
+      ScAddr const addr2 = ResolveAddr(value2);
+
+      if (addr1.IsValid())
+      {
+        ScAddr src, trg;
+        if (m_context.GetEdgeInfo(addr1, src, trg))
+        {
+          bool needCall = true;
+          if (addr0.IsValid() && (src != addr0))
+            needCall = false;
+
+          if (addr2.IsValid() && (trg != addr2))
+            needCall = false;
+
+          if (needCall)
+            fn(src, addr1, trg);
+         }        
+      }
+    }   
+  }
+
   void Iteration(size_t orderIndex, ScTemplateSearchResult & result)
   {
     size_t const constrIndex = m_template.m_searchCachedOrder[orderIndex];
@@ -226,29 +340,20 @@ public:
     /// TODO: prevent recursive search and make test for that case
 
     ScTemplateConstr3 const & constr = m_template.m_constructions[constrIndex];
-    ScTemplateItemValue const * values = constr.GetValues();
-    ScIterator3Ptr const it3 = CreateIterator(constr);
-    while (it3->Next())
+    auto const & values = constr.GetValues();
+
+    ForEachIteration(constr, [&, this](ScAddr const & addr1,
+                                       ScAddr const & addr2,
+                                       ScAddr const & addr3)
     {
-      /// check if search in structure
-      if (m_struct.IsValid())
-      {
-        if (!CheckInStruct(it3->Get(0)) ||
-            !CheckInStruct(it3->Get(1)) ||
-            !CheckInStruct(it3->Get(2)))
-        {
-          continue;
-        }
-      }
-
       // do not make cycle for optimization issues (remove comparsion expresion)
-      m_resultAddrs[resultIdx] = it3->Get(0);
-      m_resultAddrs[resultIdx + 1] = it3->Get(1);
-      m_resultAddrs[resultIdx + 2] = it3->Get(2);
+      m_resultAddrs[resultIdx] = addr1;
+      m_resultAddrs[resultIdx + 1] = addr2;
+      m_resultAddrs[resultIdx + 2] = addr3;
 
-      RefReplacement(values[0], it3->Get(0));
-      RefReplacement(values[1], it3->Get(1));
-      RefReplacement(values[2], it3->Get(2));
+      RefReplacement(values[0], addr1);
+      RefReplacement(values[1], addr2);
+      RefReplacement(values[2], addr3);
 
       if (orderIndex == finishIdx)
       {
@@ -262,7 +367,7 @@ public:
       UnrefReplacement(values[0]);
       UnrefReplacement(values[1]);
       UnrefReplacement(values[2]);
-    }
+    });
   }
 
   bool operator () (ScTemplateSearchResult & result)