+++ sc-memory/cpp/sc_utils.hpp
@@ -97,17 +97,17 @@ public:
 
   ObjectType & operator * () const
   {
-    SC_ASSERT(IsValid(), ());
+    SC_ASSERT(IsPtrValid(), ());
     return *m_object;
   }
 
   ObjectType * operator -> () const
   {
-    SC_ASSERT(IsValid(), ());
+    SC_ASSERT(IsPtrValid(), ());
     return m_object;
   }
 
-  inline bool IsValid() const
+  inline bool IsPtrValid() const
   {
     return m_object != 0;
   }
@@ -125,6 +125,7 @@ public:
   }
 
 private:
+
   void InitRef()
   {
     m_refCount = new RefCount();
@@ -197,6 +198,21 @@ namespace utils
 class StringUtils
 {
 public:
+  _SC_EXTERN static void ToLowerCase(std::string & str);
+  _SC_EXTERN static void ToUpperCase(std::string & str);
+
+  _SC_EXTERN static bool StartsWith(std::string const & str, std::string const & pattern, bool lowerCase);
+  _SC_EXTERN static bool EndsWith(std::string const & str, std::string const & pattern, bool lowerCase);
+
+  _SC_EXTERN static void SplitFilename(std::string const & qualifiedName, std::string & outBasename, std::string & outPath);
+
+  _SC_EXTERN static void TrimLeft(std::string & str);
+  _SC_EXTERN static void TrimRight(std::string & str);
+  _SC_EXTERN static void Trim(std::string & str);
+
+  _SC_EXTERN static std::string GetFileExtension(std::string const & filename);
+  _SC_EXTERN static std::string NormalizeFilePath(std::string const & init, bool makeLowerCase);
+
   _SC_EXTERN static std::string ReplaceAll(std::string const & source, std::string const & replaceWhat, std::string const & replaceWithWhat);
 };